Cleanup in layers, drop QWERTY, add NAV and rename ACCENTS to AUX
NAV & AUX are inspired by algernon, but in fact addresses the issue
with original layout: WASD and HJKL might be good nav keys without context,
but sticking to current keyboard it's better to move it a little bit,
to ensure that fingers 1-3 can be used left-down-right.
Others are somewhat intuitive changes.

Added scrolling by mouse.

I'm not sure whether the mouse buttons are at the right location, we will see.

QWERTY layer dropped as it's basically just the default firmware.
AUX is planned to be used for multimedia keys, and at that point FUNCTION
layer will be pointless. For shfit, alt, space, crtl etc. the FN version
goes into either AUX or both AUX and NAV.